+++ patterns/site/header-alt.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.1.8
+ * @version  1.2.5
  */
 
 namespace WebManDesign\Zooey\Content;
@@ -60,7 +60,7 @@ Block_Pattern::add_pattern_args( __FILE__, array(
 			</div>
 			<!-- /wp:group -->
 
-			<!-- wp:navigation {"icon":"menu","overlayBackgroundColor":"secondary","overlayTextColor":"contrast-alt","layout":{"type":"flex","justifyContent":"center"},"style":{"spacing":{"blockGap":"1.2rem"},"typography":{"fontStyle":"normal","fontWeight":"500"}},"anchor":"site-navigation"} /-->
+			<!-- wp:navigation {"icon":"menu","overlayBackgroundColor":"secondary","overlayTextColor":"contrast-alt","layout":{"type":"flex","justifyContent":"center"},"style":{"spacing":{"blockGap":"1.2rem"},"typography":{"fontStyle":"normal","fontWeight":"500"}},"anchor":"site-navigation","ariaLabel":"<?php echo esc_attr_x( 'Primary', 'Navigational menu label.', 'zooey' ); ?>"} /-->
 
 		</div>
 		<!-- /wp:group -->